requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
